change the set variable type of price from Int to Float

+++ Carpool/app/src/main/java/com/example/carpool/CarpoolPost.kt
@@ -61,7 +61,7 @@ class CarpoolPost : ParseObject () {
         put(KEY_DESCRIPTION, description)
     }
 
-    fun setPrice(price: Int) {
+    fun setPrice(price: Float) {
         put(KEY_PRICE, price)
     }
